+++ packages/implementations/react/src/composite/Modal/src/modal.types.ts
@@ -19,14 +19,15 @@ export interface ModalTyping {
    */
   open: boolean;
   /**
+   * Specifies the ID of the portal element where the modal should be rendered.
+   * This can be useful for rendering the modal in a different part of the DOM, such as a modal or overlay.
+   */
+  portalId?: string;
+  /**
    * Callback fired when the component requests to be closed.
    * @TJS-type (open: boolean) => void;
    */
   onDismiss: (open: boolean) => void;
-  /**
-   * Id to be embedded in the portal element
-   */
-  portalId?: string;
 }
 
 export type ModalProperties = ModalTyping & ModalSprinkle & ModalVariants;