+++ packages/styles/src/packages/atomic/select/rarui-select.types.ts
@@ -1,9 +1,11 @@
 import { RecipeVariants } from "@vanilla-extract/recipes";
 import { selectStyles } from "./rarui-select.css";
+import { DropdownSprinkle } from "../../composite/dropdown/rarui-dropdown.types";
 import { StandardLonghandProperties } from "../../../index.types";
 
 export type SelectVariants = NonNullable<
   RecipeVariants<typeof selectStyles.select>
 >;
 
-export type SelectSprinkle = Pick<StandardLonghandProperties, "maxHeight">;
+export type SelectSprinkle = Pick<StandardLonghandProperties, "maxHeight"> &
+  Pick<DropdownSprinkle, "zIndex">;